+++ common/src/main/java/dev/ryanhcode/sable/api/sublevel/SubLevelContainer.java
@@ -11,21 +11,21 @@ import dev.ryanhcode.sable.sublevel.plot.PlotChunkHolder;
 import dev.ryanhcode.sable.sublevel.storage.SubLevelOccupancySavedData;
 import dev.ryanhcode.sable.sublevel.storage.SubLevelRemovalReason;
 import dev.ryanhcode.sable.util.iterator.ListBackedFilterIterator;
-import it.unimi.dsi.fastutil.objects.ObjectArrayList;
-import it.unimi.dsi.fastutil.objects.ObjectList;
-import it.unimi.dsi.fastutil.objects.ObjectLists;
+import it.unimi.dsi.fastutil.objects.*;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.SectionPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.players.PlayerList;
+import net.minecraft.util.Mth;
 import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.chunk.LevelChunk;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.Nullable;
 import org.joml.Vector2i;
+import org.joml.Vector3dc;
 
 import java.util.*;
 
@@ -49,11 +49,11 @@ public abstract class SubLevelContainer {
     /**
      * All of the loaded sub-levels in the plotgrid
      */
-    private final List<SubLevel> allSubLevels = new ObjectArrayList<>();
+    private final List<SubLevel> allSubLevels = new ReferenceArrayList<>();
     /**
      * All of the loaded sub-levels in the plotgrid, by uuid
      */
-    private final Map<UUID, SubLevel> subLevelsByUUID = new HashMap<>();
+    private final Map<UUID, SubLevel> subLevelsByUUID = new Object2ObjectOpenHashMap<>();
     /**
      * The occupancy of the plotgrid, including loaded and unloaded plots
      */
@@ -61,7 +61,7 @@ public abstract class SubLevelContainer {
     /**
      * All observers/listeners for the plotgrid
      */
-    private final List<SubLevelObserver> observers = new ObjectArrayList<>();
+    private final List<SubLevelObserver> observers = new ReferenceArrayList<>();
 
     /**
      * The level of the plotgrid
@@ -358,6 +358,13 @@ public abstract class SubLevelContainer {
      */
     public boolean inBounds(final BlockPos pos) {
         return this.inBounds(pos.getX() >> SectionPos.SECTION_BITS, pos.getZ() >> SectionPos.SECTION_BITS);
+    }
+
+    /**
+     * @return if a global position is within the plotgrid.
+     */
+    public boolean inBounds(final Vector3dc pos) {
+        return this.inBounds(Mth.floor(pos.x()) >> SectionPos.SECTION_BITS, Mth.floor(pos.z()) >> SectionPos.SECTION_BITS);
     }
 
     /**