+++ fabric/src/main/java/dev/ryanhcode/sable/fabric/platform/SableLoaderPlatformImpl.java
@@ -2,8 +2,6 @@ package dev.ryanhcode.sable.fabric.platform;
 
 import dev.ryanhcode.sable.platform.SableLoaderPlatform;
 import net.fabricmc.loader.api.FabricLoader;
-import org.jetbrains.annotations.Nullable;
-
 import java.nio.file.Path;
 
 public class SableLoaderPlatformImpl implements SableLoaderPlatform {
@@ -18,7 +16,6 @@ public class SableLoaderPlatformImpl implements SableLoaderPlatform {
 	}
 
 	@Override
-	@Nullable
 	public Path getGameDirectory() {
 		return FabricLoader.getInstance().getGameDir();
 	}