+++ changelog.md
@@ -1,2 +1,2 @@
- - Fix issues with glibc support on Linux Sable Rapier natives
- - Fix Create Blueprints not working on sub-levels
+- Fix a crash with sub-level splitting
+- Fix and add recovery for an issue with sub-levels not loading in because their bounds got set to [0,0,0,0,0,0]